ealabsservicebackend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.31
pydantic==2.8.2
python-multipart==0.0.9
reportlab==4.2.0
python-jose[cryptography]
passlib[bcrypt]
